Rework benchmarks
