+++ collect_app/src/androidTest/java/org/odk/collect/android/benchmark/EntitiesBenchmarkTest.kt
@@ -18,18 +18,8 @@ import org.odk.collect.android.support.pages.MainMenuPage
 import org.odk.collect.android.support.rules.CollectTestRule
 import org.odk.collect.android.support.rules.TestRuleChain.chain
 import org.odk.collect.android.test.BuildConfig.ENTITIES_FILTER_PROJECT_URL
+import org.odk.collect.android.test.BuildConfig.THOUSAND_MEDIA_FILE_ENTITY_LIST_PROJECT_URL
 import org.odk.collect.strings.R
-
-/**
- * Benchmarks the performance of entity follow up forms. [ENTITIES_FILTER_PROJECT_URL] should
- * be set to a project that contains the "100k Entities Filter" benchmark form and the
- * "entities_100k" entity list.
- *
- * Devices that currently pass:
- * - Fairphone 3
- * - Pixel 3
- *
- */
 
 @RunWith(AndroidJUnit4::class)
 class EntitiesBenchmarkTest {
@@ -39,8 +29,18 @@ class EntitiesBenchmarkTest {
     @get:Rule
     var chain: RuleChain = chain(TestDependencies(true)).around(rule)
 
+    /**
+     * Benchmarks the performance of entity follow up forms. [ENTITIES_FILTER_PROJECT_URL] should
+     * be set to a project that contains the "100k Entities Filter" benchmark form and the
+     * "entities_100k" entity list.
+     *
+     * Devices that currently pass:
+     * - Fairphone 3
+     * - Pixel 3
+     *
+     */
     @Test
-    fun run() {
+    fun oneHundredThousandEntities() {
         assertThat(
             "Need to set ENTITIES_FILTER_PROJECT_URL before running!",
             ENTITIES_FILTER_PROJECT_URL,
@@ -94,6 +94,48 @@ class EntitiesBenchmarkTest {
             .answerQuestion("Which value do you want to filter by?", "1024")
             .benchmark("Filtering select", 3, benchmarker) {
                 it.swipeToNextQuestion("Filtered select")
+            }
+
+        benchmarker.assertResults()
+    }
+
+    /**
+     * Benchmarks the performance of updating forms with entity lists and many media files.
+     * [THOUSAND_MEDIA_FILE_ENTITY_LIST_PROJECT_URL] should be set to a project that contains the
+     * "1000-media-files-entity-list" form.
+     *
+     * Devices that currently pass:
+     * - Fairphone 3
+     */
+    @Test
+    fun oneThousandMediaFilesWithEntityList() {
+        assertThat(
+            "Need to set THOUSAND_MEDIA_FILE_ENTITY_LIST_PROJECT_URL before running!",
+            THOUSAND_MEDIA_FILE_ENTITY_LIST_PROJECT_URL,
+            not(blankOrNullString())
+        )
+
+        val benchmarker = Benchmarker()
+
+        rule.startAtFirstLaunch()
+            .clickManuallyEnterProjectDetails()
+            .inputUrl(THOUSAND_MEDIA_FILE_ENTITY_LIST_PROJECT_URL)
+            .addProject()
+
+            // Download all forms
+            .clickGetBlankForm()
+            .clickGetSelected()
+            .clickOKOnDialog(MainMenuPage())
+
+            .clickGetBlankForm()
+            .benchmark(
+                "Redownloading a form with 1k media files and entity list when there are no updates",
+                15,
+                benchmarker
+            ) {
+                it
+                    .clickGetSelected()
+                    .clickOKOnDialog(MainMenuPage())
             }
 
         benchmarker.assertResults()
