+++ collect_app/src/androidTest/java/org/odk/collect/android/benchmark/FormsUpdateBenchmarkTest.kt
@@ -14,7 +14,6 @@ import org.odk.collect.android.support.TestDependencies
 import org.odk.collect.android.support.pages.MainMenuPage
 import org.odk.collect.android.support.rules.CollectTestRule
 import org.odk.collect.android.support.rules.TestRuleChain.chain
-import org.odk.collect.android.test.BuildConfig.THOUSAND_MEDIA_FILE_ENTITY_LIST_PROJECT_URL
 import org.odk.collect.android.test.BuildConfig.THOUSAND_MEDIA_FILE_PROJECT_URL
 
 @RunWith(AndroidJUnit4::class)
@@ -25,7 +24,7 @@ class FormsUpdateBenchmarkTest {
     val chain: RuleChain = chain(TestDependencies(true)).around(rule)
 
     /**
-     * Benchmarks the performance of updating forms. [THOUSAND_MEDIA_FILE_PROJECT_URL] should
+     * Benchmarks the performance of check for updates. [THOUSAND_MEDIA_FILE_PROJECT_URL] should
      * be set to a project that contains the "1000-media-files" benchmark form.
      *
      * Devices that currently pass:
@@ -59,57 +58,6 @@ class FormsUpdateBenchmarkTest {
                 it
                     .clickGetBlankForm()
                     .clickSelectAll()
-            }
-
-            .benchmark(
-                "Redownloading a form with 1k media files when there are no updates",
-                12,
-                benchmarker
-            ) {
-                it
-                    .clickGetSelected()
-                    .clickOKOnDialog(MainMenuPage())
-            }
-
-        benchmarker.assertResults()
-    }
-
-    /**
-     * Benchmarks the performance of updating forms. [THOUSAND_MEDIA_FILE_ENTITY_LIST_PROJECT_URL] should
-     * be set to a project that contains the "1000-media-files-entity-list" form.
-     *
-     * Devices that currently pass:
-     * - Fairphone 3
-     */
-    @Test
-    fun oneThousandMediaFilesWithEntityList() {
-        assertThat(
-            "Need to set THOUSAND_MEDIA_FILE_ENTITY_LIST_PROJECT_URL before running!",
-            THOUSAND_MEDIA_FILE_ENTITY_LIST_PROJECT_URL,
-            not(blankOrNullString())
-        )
-
-        val benchmarker = Benchmarker()
-
-        rule.startAtFirstLaunch()
-            .clickManuallyEnterProjectDetails()
-            .inputUrl(THOUSAND_MEDIA_FILE_ENTITY_LIST_PROJECT_URL)
-            .addProject()
-
-            // Download all forms
-            .clickGetBlankForm()
-            .clickGetSelected()
-            .clickOKOnDialog(MainMenuPage())
-
-            .clickGetBlankForm()
-            .benchmark(
-                "Redownloading a form with 1k media files and entity list when there are no updates",
-                15,
-                benchmarker
-            ) {
-                it
-                    .clickGetSelected()
-                    .clickOKOnDialog(MainMenuPage())
             }
 
         benchmarker.assertResults()
